requirements.txt
flask
black
msal
requests
python-dotenv
langchain
langchain-core
langchain-community
bs4